- implement auth flow

+++ src/tasks/tasks.service.ts
@@ -1,25 +1,33 @@
 import {Injectable, NotFoundException} from '@nestjs/common';
-import { TaskStatus} from "./interfaces";
+import {TaskStatus} from "./interfaces";
 import {CreateTaskDto} from "./dto/create-task.dto";
-import { v4 as uuid } from 'uuid';
+import {v4 as uuid} from 'uuid';
 import {GetTasksFilterDto} from "./dto/get-tasks-filter.dto";
 import {InjectRepository} from "@nestjs/typeorm";
 import {TaskRepository} from "./tasks.repository";
-import {Task} from "./tasks.entity";
+import {Task} from "./task.entity";
+import {User} from "../auth/user.entity";
 
 @Injectable()
 export class TasksService {
     constructor(
         @InjectRepository(TaskRepository)
         private taskRepository: TaskRepository,
-    ) {}
-
-    async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-        return this.taskRepository.getTasks(filterDto);
+    ) {
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
+    async getTasks(
+        filterDto: GetTasksFilterDto,
+        user: User,
+    ): Promise<Task[]> {
+        return this.taskRepository.getTasks(filterDto, user);
+    }
+
+    async getTaskById(
+        id: number,
+        user: User,
+    ): Promise<Task> {
+        const found = await this.taskRepository.findOne({where: {id, userId: user.id}});
 
         if (!found) {
             throw new NotFoundException(`Task with ID "${id}" not found`);
@@ -28,20 +36,30 @@ export class TasksService {
         return found;
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return this.taskRepository.createTask(createTaskDto);
+    async createTask(
+        createTaskDto: CreateTaskDto,
+        user: User,
+    ): Promise<Task> {
+        return this.taskRepository.createTask(createTaskDto, user);
     }
 
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(
+        id: number,
+        user: User,
+    ): Promise<void> {
+        const result = await this.taskRepository.delete({id, userId: user.id});
 
         if (result.affected === 0) {
             throw new NotFoundException(`Task with ID "${id}" not found`);
         }
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(
+        id: number,
+        status: TaskStatus,
+        user: User,
+    ): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
         return task;
